+++ ChangeLog.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Refactored `Show-FolderUsage` to support non-Windows platforms in PowerShell 7.
+
+### Fixed
+- Fixed localization error with Invariant cultures.
+- Fixed error in exporting module functions.
 
 ## [0.11.0] - 2025-06-13